requirements.txt
absl-py==0.9.0
appdirs==1.4.4
appnope==0.1.0
asn1crypto==1.3.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
boto3==1.12.41
botocore==1.15.41
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.4.1
conda==4.8.3
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
dask==2.17.2
decorator==4.4.2
defusedxml==0.6.0
detecta==0.0.5
distlib==0.3.0
distributed==2.17.0
docutils==0.15.2
elasticsearch==7.7.1
entrypoints==0.3
et-xmlfile==1.0.1
EZSheets==0.0.6
featuretools==0.15.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.12.0
Flask-SQLAlchemy==2.4.3
FlowCal==1.2.2
future==0.18.2
gast==0.3.3
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
HeapDict==1.0.1
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.5
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1589812474002/work
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lightgbm==2.3.1
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.0
matplotlib==3.2.1
mistune==0.8.4
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nltk==3.5
notebook==6.0.3
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1588605093274/work
oauthlib==3.1.0
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.3
palettable==3.3.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycorenlp==0.3.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-docx==0.8.10
pytorch-pretrained-bert==0.6.2
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.5.14
reppy==0.4.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel-yaml==0.15.87
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
seqeval==0.0.12
six==1.14.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.17
tblib==1.6.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
torch==1.5.0
torchtext==0.1.1
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
urllib3==1.25.9
virtualenv==20.0.21
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.2.1
xgboost==1.1.0
xlrd==1.2.0
XlsxWriter==1.2.8
zict==2.0.0
zipp==3.1.0
